beginning to separate out firewall rule from connection dialog
